Rx java usage
The storage write now returns an Observabe<CacheEntry> instead
of just a CacheEntry. This was the executor could be skipped
and just subscribe on a io thread instead. Also the file exceptions
from the storage could be wrapped into the observable
remove messy catch code.

Changed the syncronization guard to a read write lock instead.